+++ tartufo/scanner.py
@@ -28,7 +28,13 @@ import git
 import pygit2
 
 from tartufo import config, types, util
-from tartufo.types import BranchNotFoundException, Rule, TartufoException
+from tartufo.types import (
+    BranchNotFoundException,
+    Rule,
+    TartufoException,
+    MatchType,
+    Scope,
+)
 
 BASE64_REGEX = re.compile(r"[A-Z0-9=+/_-]+", re.IGNORECASE)
 HEX_REGEX = re.compile(r"[0-9A-F]+", re.IGNORECASE)
@@ -382,18 +388,25 @@ class ScannerBase(abc.ABC):  # pylint: disable=too-many-instance-attributes
 
         :param rule: Rule to perform match
         :param string: string to match against rule pattern
+        :param line: Source line containing string of interest
         :param path: path to match against rule path_pattern
         :return: True if string and path matched, False otherwise.
         """
         match = False
-        if rule.re_match_type == "match":
+        if rule.re_match_scope == Scope.Word:
+            scope = string
+        elif rule.re_match_scope == Scope.Line:
+            scope = line
+        else:
+            raise TartufoException(f"Invalid value for scope: {rule.re_match_scope}")
+        if rule.re_match_type == MatchType.Match:
             if rule.pattern:
-                match = rule.pattern.match(string) is not None
+                match = rule.pattern.match(scope) is not None
             if rule.path_pattern:
                 match = match and rule.path_pattern.match(path) is not None
-        elif rule.re_match_type == "search":
+        elif rule.re_match_type == MatchType.Search:
             if rule.pattern:
-                match = rule.pattern.search(line) is not None
+                match = rule.pattern.search(scope) is not None
             if rule.path_pattern:
                 match = match and rule.path_pattern.search(path) is not None
 
@@ -403,6 +416,7 @@ class ScannerBase(abc.ABC):  # pylint: disable=too-many-instance-attributes
         """Find whether the signature of some data has been excluded in configuration.
 
         :param string: String to check against rule pattern
+        :param line: Source line containing string of interest
         :param path: Path to check against rule path pattern
         :return: True if excluded, False otherwise
         """